requirements.txt
openai>=2.0.0
rich>=13.7.0
python-dotenv>=1.0.1
requests>=2.31.0
ddgs>=9.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
playwright>=1.40.0
dnspython>=2.4.0
